requirements.txt
faiss-cpu
langchain
openai
python-dotenv
fastapi
jinja2
python-multipart
uvicorn
tiktoken
langchain_community
langchain_openai